Separate `httpAPI` package into `router` and `controller` packages
