+++ httpAPI/router/api.go
@@ -1,4 +1,4 @@
-package httpAPI
+package router
 
 import (
 	"log"
